fixed most bugs found by Fingbugs. Only two categories of bugs are not fixed (execute sql from nonconstant string, and use of float instead of double).
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonWebsite/trunk@48688 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Model/src/java/org/apidb/apicommon/model/UserFileFactory.java
@@ -114,6 +114,8 @@ public class UserFileFactory {
                 userFile.setFileSize(fileOnDisk.length());
 
                 insertUserFileMetaData(userFile);
+            } else {
+              throw new UserFileUploadException("File name not specified.");
             }
         } catch (IOException ioe) {
             String msg = "Could not write '" + fileName + "' to '" + filePath
@@ -122,11 +124,13 @@ public class UserFileFactory {
             throw new UserFileUploadException(msg + "\n" + ioe);
         } catch (Exception e) {
             logger.warn(e);
+            if (fileOnDisk != null) {
             logger.warn("Deleting " + fileOnDisk.getPath());
             if (fileOnDisk.exists() && !fileOnDisk.delete())
                 logger.warn("\nUnable to delete "
                         + fileOnDisk.getPath()
                         + ". This file may not be correctly recorded in the database.");
+            }
             throw new UserFileUploadException(e);
         }
